refactor: relocate target (API)
